requirements.txt
albumentations
opencv-python
numpy
yacs
Pillow
cython
tqdm
PyYAML
